requirements.txt
openai>=1.30
python-dotenv>=1.0
pydantic>=2.6
pyyaml>=6.0
pymysql>=1.1
httpx>=0.27
fastapi>=0.110
uvicorn>=0.29
sentence-transformers>=2.7
numpy>=1.26
